requirements.txt
annotated-types==0.5.0
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
copier==8.3.0
decorator==5.1.1
dunamai==1.18.1
exceptiongroup==1.1.3
flet==0.10.3
flet-core==0.10.3
flet-runtime==0.10.3
funcy==2.0
future==0.18.3
geocoder==1.38.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
plumbum==1.8.2
prompt-toolkit==3.0.36
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pypng==0.20220715.0
python-dotenv==1.0.0
PyYAML==6.0.1
pyyaml-include==1.3.1
qrcode==7.4.2
questionary==2.0.1
ratelim==0.1.6
repath==0.9.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
typing_extensions==4.8.0
urllib3==2.0.6
watchdog==3.0.0
wcwidth==0.2.8
websocket-client==1.6.3
websockets==11.0.3
